# api/openapi/todolist.yaml
openapi: 3.0.3
info:
  title: ToDoList
  description: backend for TodoIst
  version: 1.0.0
servers:
  - url: 'https://{hostname}/api'
    variables:
      hostname:
        default: localhost
paths:
  /group/create:
    post:
      operationId: createGroup
      requestBody:
        description: create new a group
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          headers:
            Content-Location:
              description: created a group
              schema:
                type: string
          description: group created
        '400':
          description: invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /group/{groupId}:
    get:
      operationId: getGroup
      parameters:
        - in: path
          name: groupId
          schema:
            type: string
            format: uuid
          required: true
          description: group id
      responses:
        '200':
          description: found course
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GroupResponse'

  /group/{groupId}/addTask:
    post:
      operationId: AddTaskInGroup
      parameters:
        - in: path
          name: groupId
          schema:
            type: string
            format: uuid
          required: true
          description: group id
      requestBody:
        description: create new a task
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          headers:
            Content-Location:
              description: task created and added in group
              schema:
                type: string
          description: task created

  /task/change:
    post:
      operationId: ChangeTask
      requestBody:
        description: change a task
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Task'
      responses:
        '201':
          description: task changed


components:
  schemas:
    CreateGroupRequest:
      type: object
      required: [ title, description ]
      properties:
        title:
          type: string
        description:
          type: string

    GroupResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        taskIDs:
          type: array
          items:
            type: string
            format: uuid

    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: integer
        createDate:
          type: string
        currentDoingDate:
          type: string
        endDate:
          type: string




    Error:
      type: object
      required: [ slug, details ]
      properties:
        slug:
          type: string
        details:
          type: string